Initial import.
git-svn-id: https://s.snth.net/svn/neverball/trunk@1 78b8d119-cf0a-0410-b17c-f493084dd1d7

+++ putt/game.c
@@ -30,7 +30,7 @@
 /*---------------------------------------------------------------------------*/
 
 static struct s_file file;
-static short         ball;
+static int           ball;
 
 static float view_a;                    /* Ideal view rotation about Y axis  */
 static float view_m;
